+++ src/Symfony/Bridge/Twig/Tests/Extension/Fixtures/StubTranslator.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Bridge\Twig\Tests\Extension\Fixtures;
 
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class StubTranslator implements TranslatorInterface
 {